orchestrator/email_coordinator/draft_manager/draft_composer/requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
pyyaml>=6.0.0
